Basic Multiplayer
Got the basic multiplayer working with a Client/Dedicated server style of communication. I plan on having the server do all of the calculations, then send that information to the clients so that the game is consistent amongst all clients, and possibly to prevent cheating? Don't really know if that last one is important at all.